Instead of always getting a new VPN config file from the server, first try to use the existing config file and otherwise use the new server route to renew the keypair

+++ vantage6-node/vantage6/node/__init__.py
@@ -478,18 +478,23 @@ class Node(object):
         VPNManager
             Manages the VPN connection
         """
-        # get the ovpn configuration from the server
-        success, ovpn_config = self.server_io.get_vpn_config()
-        if not success:
-            self.log.warn("Obtaining VPN configuration file not successful!")
-            self.log.warn("Disabling node-to-node communication via VPN")
-            self.has_vpn = False
-            return
-
-        # write ovpn config to node docker volume
         ovpn_file = os.path.join(self.ctx.data_dir, VPN_CONFIG_FILE)
-        with open(ovpn_file, 'w') as f:
-            f.write(ovpn_config)
+
+        # if vpn config doesn't exist, get it and write to disk
+        if not os.path.isfile(ovpn_file):
+            # get the ovpn configuration from the server
+            success, ovpn_config = self.server_io.get_vpn_config()
+            if not success:
+                self.log.warn(
+                    "Obtaining VPN configuration file not successful!")
+                self.log.warn("Disabling node-to-node communication via VPN")
+                return
+
+            # write ovpn config to node docker volume
+            with open(ovpn_file, 'w') as f:
+                f.write(ovpn_config)
+        else:
+            self.log.debug("Using existing VPN configuration file")
 
         # set up the VPN connection via docker containers
         self.log.debug("Setting up VPN client container")
@@ -497,7 +502,19 @@ class Node(object):
             isolated_network_mgr=isolated_network_mgr,
             node_name=self.ctx.name
         )
-        vpn_manager.connect_vpn(ovpn_file=ovpn_file)
+        try:
+            vpn_manager.connect_vpn(ovpn_file=ovpn_file)
+        except Exception:
+            self.log.debug(
+                "Could not connect to VPN. Trying to refresh keypair...")
+            # Connecting VPN failed, which may be due to an expired keypair.
+            # Refresh the client's keypair and try to connect again
+            success = self.server_io.refresh_vpn_keypair(ovpn_file=ovpn_file)
+            if not success:
+                self.log.warn("Refreshing VPN keypair not successful!")
+                self.log.warn("Disabling node-to-node communication via VPN")
+                return
+            vpn_manager.connect_vpn(ovpn_file=ovpn_file)
         return vpn_manager
 
     def connect_to_socket(self):
